move compute_base_fee test to base_fee module
